% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{apply_redcap_form_labels}
\alias{apply_redcap_form_labels}
\title{Apply REDCap metadata labels to a dataset}
\usage{
apply_redcap_form_labels(
  data,
  metadata,
  checkbox = c("both", "label", "collapse"),
  keep_checkbox_columns = TRUE,
  checkbox_sep = "; "
)
}
\arguments{
\item{data}{REDCap records.}

\item{metadata}{REDCap metadata.}

\item{checkbox}{One of \code{"label"}, \code{"collapse"}, or \code{"both"}.}

\item{keep_checkbox_columns}{If \code{FALSE}, remove checkbox child columns after collapse.}

\item{checkbox_sep}{Separator when collapsing checkbox selections.}
}
\value{
A labelled data frame.
}
\description{
Applies:
\itemize{
\item variable labels
\item value labels for non-checkbox fields
\item checkbox labels
\item optional checkbox collapsing
}
}
