% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{build_redcap_form_col_types}
\alias{build_redcap_form_col_types}
\title{Build a readr column specification from REDCap metadata}
\usage{
build_redcap_form_col_types(field_metadata, form_field_names)
}
\arguments{
\item{field_metadata}{REDCap form metadata tibble}

\item{form_field_names}{The REDCap form fields}
}
\value{
A readr cols specification.
}
\description{
Build a readr column specification from REDCap metadata
}
