% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/types.R
\name{convert_redcap_form_types}
\alias{convert_redcap_form_types}
\title{Convert REDCap export columns using metadata-driven types}
\usage{
convert_redcap_form_types(data, field_metadata, na = c("", "NA", "UNK"))
}
\arguments{
\item{data}{A data frame, ideally with all columns as character.}

\item{field_metadata}{REDCap field name metadata tibble.}

\item{na}{Character vector of strings to treat as missing.}
}
\value{
A typed tibble.
}
\description{
Convert REDCap export columns using metadata-driven types
}
