% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{export_events}
\alias{export_events}
\title{Export the events for a project}
\usage{
export_events(client)
}
\arguments{
\item{client}{A \code{redcap_client}.}
}
\value{
A tibble
}
\description{
This method allows you to export the events for a project
}
