% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{export_instrument_event_mappings}
\alias{export_instrument_event_mappings}
\title{Export the instrument-event mappings for a project}
\usage{
export_instrument_event_mappings(client)
}
\arguments{
\item{client}{A \code{redcap_client}.}
}
\value{
A tibble
}
\description{
This method allows you to export the instrument-event mappings for a project
(i.e., how the data collection instruments are designated for certain events in a longitudinal project).
}
