% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{export_instruments}
\alias{export_instruments}
\title{Export the data collection instruments for a project}
\usage{
export_instruments(client)
}
\arguments{
\item{client}{A \code{redcap_client}.}
}
\value{
A tibble
}
\description{
This method allows you to export a list of the data collection instruments for a project.
This includes their unique instrument name as seen in the second column of the Data Dictionary,
as well as each instrument's corresponding instrument label, which is seen on a project's left-hand
menu when entering data. The instruments will be ordered according to their order in the project.
}
