% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{export_metadata}
\alias{export_metadata}
\title{Export the metadata for a project}
\usage{
export_metadata(client)
}
\arguments{
\item{client}{A \code{redcap_client}.}
}
\value{
A tibble of REDCap metadata
}
\description{
This method allows you to export the metadata for a project
}
