% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{export_project_info}
\alias{export_project_info}
\title{Export the project info}
\usage{
export_project_info(client)
}
\arguments{
\item{client}{A \code{redcap_client}.}
}
\value{
A tibble
}
\description{
This method allows you to export some of the basic attributes of a given REDCap project,
such as the project's title, if it is longitudinal, if surveys are enabled,
the time the project was created and moved to production, etc.
}
