% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{export_records}
\alias{export_records}
\title{Export the metadata for a project}
\usage{
export_records(client, forms, fields, events, raw_or_label = c("raw", "label"))
}
\arguments{
\item{client}{A \code{redcap_client}.}

\item{forms}{A comma separated string giving the forms to export}

\item{fields}{A comma separated string giving the fields to export}

\item{events}{A comma separated string giving the events to export}

\item{raw_or_label}{Return raw or labelled field values}
}
\value{
A tibble
}
\description{
This method allows you to export the metadata for a project
}
