% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{export_redcap}
\alias{export_redcap}
\title{Export all REDCap forms along with metadata}
\usage{
export_redcap(client, raw_or_label = "raw")
}
\arguments{
\item{client}{REDCap client}

\item{raw_or_label}{Export 'raw' or 'label' data from REDCap.}
}
\value{
A list of REDCap data.
}
\description{
The aim of this function is to export all of the relevant REDCap data.
In particular, we want:
\itemize{
\item the raw REDCap records
\item the data dictionary and other metadata necessary to determine field types and labels.
}
}
