% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{get_value_labels}
\alias{get_value_labels}
\title{Get value labels from REDCap metadata}
\usage{
get_value_labels(data, value_labels_dict)
}
\arguments{
\item{metadata}{REDCap metadata tibble.}
}
\value{
Named list of value label vectors.
}
\description{
Excludes checkbox fields because checkbox labels are handled separately.
}
