% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labeling.R
\name{get_variable_labels}
\alias{get_variable_labels}
\title{Get variable labels from REDCap metadata}
\usage{
get_variable_labels(metadata)
}
\arguments{
\item{metadata}{REDCap metadata tibble.}
}
\value{
Named character vector of variable labels.
}
\description{
Get variable labels from REDCap metadata
}
