% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{print.redcap_client}
\alias{print.redcap_client}
\title{Print a REDCap API client}
\usage{
\method{print}{redcap_client}(x, ...)
}
\arguments{
\item{x}{A REDCap API client.}

\item{...}{Other arguments}
}
\description{
Print a REDCap API client
}
