% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client.R
\name{redcap_client}
\alias{redcap_client}
\title{Create a REDCap API client}
\usage{
redcap_client(url, token)
}
\arguments{
\item{url}{REDCap API URL.}

\item{token}{REDCap API token.}
}
\value{
An object of class \code{redcap_client}.
}
\description{
Create a REDCap API client
}
