% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{save_redcap_bundle}
\alias{save_redcap_bundle}
\title{Save a REDCap bundle to disk}
\usage{
save_redcap_bundle(data, metadata, path)
}
\arguments{
\item{data}{Labelled data frame.}

\item{metadata}{REDCap metadata.}

\item{path}{Output directory.}
}
\value{
Invisibly returns \code{path}.
}
\description{
Saves:
\itemize{
\item labelled data as RDS
\item metadata as RDS and JSON
\item checkbox dictionary as CSV
}
}
