% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{split_metadata_by_form}
\alias{split_metadata_by_form}
\title{Helper function to split metadata into form-specific tibbles}
\usage{
split_metadata_by_form(metadata)
}
\arguments{
\item{metadata}{REDCap metadata tibble}
}
\value{
A list of tibbles by form
}
\description{
Helper function to split metadata into form-specific tibbles
}
